+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - If sparse feature values present on multiple servers, then only one will be returned with source picked randomly.
 
+### Removed
+
+- Remove ALL_NODE_TYPE, ALL_EDGE_TYPE, __len__ and __iter__ from Graph API.
+
 ## [0.1.57] - 2022-12-15
 
 ### Changed